requirements.txt
streamlit>=1.37
python-dotenv>=1.0
sentence-transformers>=2.7
pinecone>=5.0.0
cohere>=5.5
groq>=0.9
numpy>=1.26
pytest>=8.2
python-dotenv>=1.0
PyPDF2>=3.0.0